ATT_FUNCTS
Mapped all of the Attack Functions (Only the Special and the Physical Moves) and Wrote in the hardcodes functions with a string return type for now just to show when they are called. They will remain like this until they are all coded otherwise. Changing the return type to void will be one of the last things to alter.