Delayed Text and Stat Alteration Setters
I added a class to alter the way text is displayed to the user. It does not accept functions like "getters" or stream manipulators.
I also added a few setters commented out in the Pokemon class for a way to alter the stats of the pokemon objects in Battle. They are not final though.